Fix :: error on renaming environment names while changing cases
